Partial refactor - Move view components to its own path - Resize octicons to 167px - New git toggle!
