+++ src/touchbar/components/FoldCodePopover.js
@@ -29,16 +29,16 @@ export default class FoldCodePopover extends Component {
     const foldPromise = createOcticonImage({
       icon: 'fold',
       color: whiteColor,
-      height: 204,
-      width: 204,
+      height: 167,
+      width: 167,
       scaleFactor: 10.0,
     });
 
     const unfoldPromise = createOcticonImage({
       icon: 'unfold',
       color: whiteColor,
-      height: 204,
-      width: 204,
+      height: 167,
+      width: 167,
       scaleFactor: 10.0,
     });
