+++ src/touchbar/components/ToggleGitTabButton.js
@@ -1,17 +1,18 @@
 'use babel';
 
 import React, { Component } from 'react';
+import PropTypes from 'prop-types';
 import {
   createOcticonImage,
   executeAtomCommand,
 } from '../../utils';
 
-class SearchFilesPopover extends Component {
+class ToggleGitTabButton extends Component {
   constructor(props) {
     super(props);
 
     this.state = {
-      iconFolder: null,
+      icon: null,
     };
 
     this.buildIcon();
@@ -19,35 +20,36 @@ class SearchFilesPopover extends Component {
 
   async buildIcon() {
     const icon = await createOcticonImage({
-      icon: 'file-directory',
+      icon: 'git-commit',
       color: '#ffffff',
       height: 167,
       width: 167,
       scaleFactor: 10.0,
     });
+
     this.setState({ icon });
   }
 
   render() {
-    const { icon } = this.state;
-
     return (
-      <popover
-        icon={icon}
-      >
       <button
-        onClick={() => executeAtomCommand('fuzzy-finder:toggle-file-finder')}
+        {...this.props}
+        iconPosition="left"
+        icon={this.state.icon}
+        onClick={() => executeAtomCommand('github:toggle-git-tab')}
       >
-        Search all files
+        Git
       </button>
-        <button
-          onClick={() => executeAtomCommand('fuzzy-finder:toggle-git-status-finder')}
-        >
-          Search modified files
-        </button>
-      </popover>
     );
   }
 }
 
-export default SearchFilesPopover;
+ToggleGitTabButton.defaultProps = {
+  iconColor: '#ffffff',
+};
+
+ToggleGitTabButton.propTypes = {
+  iconColor: PropTypes.string,
+};
+
+export default ToggleGitTabButton;
